Updated openstack/openstack Project: openstack/diskimage-builder  b4266001d07b42dd78baeff18b55f5d8277a356c
Spec for changing the block device handling: partitioning

During the creation of a disk image (e.g. for a VM), there is the need
to create, setup, configure and afterwards detach some kind of storage
where the newly installed OS can be copied to or directly installed
in.

This document describes the details of partitioning.

Change-Id: I72d8459e81d7da7a8feaebbf295d6c436a686a07
Signed-off-by: Andreas Florath <andreas@florath.net>