Updated openstack/openstack Project: openstack/nova  c1a249efed3c66c85e6d47383f3e37248ca80e40
Fix root_device_name for Xen

When running the Tempest scenario test_volume_boot_pattern.\
     TestVolumeBootPattern.test_volume_boot_pattern with a
libvirt+Xen environment, the tests fails because the
root_device_name is set to 'vda' which is invalid for Xen.
When getting the BDM for the root block device, check that
the root_device_name is compatible with the used disk bus and
if it is not, adjust the root_device_name.

Change-Id: I4c231b5ef646a067be1566653b6ee89ccbee60f3
Closes-Bug: #1443898