Updated openstack/openstack Project: openstack/barbican  6912f2b4aa6ac632591415b884f8e385cfb66b30
Split serial and parallel functional test runs

The functional tests as written take almost an hour to run.  This long run
is on the border of timing out and failing the test due time.  The
tests are run twice: once in serial and then a second time in parallel.
While an added bonus to catch race conditions, the second run is proving
to be a burden when running in the gate, by bringing about the need for
frequent rechecks and reverifies.

This patch splits the test runs into two parts: serial and parallel.
For now, we'll only run the serial tests in the gate.

Change-Id: Ife616ed912cdaa43ef70ea99b320085acd1827e0