Updated openstack/openstack Project: openstack/nova  918d1b85fcf98a011f298ae318c389d84fa8ba1e
Fresh resource provider in RT must have generation 0

When the resource tracker creates a new resource provider it sets
the generation to 1. This is wrong, because on the server side the
default value for a generation, and thus the generation for a new
provider is 0. This fixes that situation and thus avoids causing the
first inventory write operation from the resource tracker to the
placement API being a 409 conflict.

Change-Id: Iaa9e786f33049799678f21b5dfaac417c6601ac8
Closes-Bug: #1660682