Updated openstack/openstack Project: openstack/cinder  655010356280eedbbb3e3b4b93bf32a814f8c0b5
clean up replication volume on an xiv backend

in the xiv portion of the ibm_storage driver, delete volume
silently failed if the volume was replicated, due to
invalid argument name.
The result was an unmanaged volume left on the storage,
taking up space.
The argument name has been fixed and the volume is cleaned up.

Change-Id: I58bc8f91bdee713dc064547e682f95d9b2cb6cd2
closes-bug: 1660127