Updated openstack/openstack Project: openstack/glance  a77ca91133949559860a364f38237906571655bb
Refine migration query added with CI change

The query added with CI change [0] was using the python 'and'
instead of sqlalchemy's 'and_()'. This patch changes that.
This patch also further refines the query to look specifically
for visility 'private' instead of '<> public' for accuracy during
setting of visibility to 'shared'.

[0] I94bc7708b291ce37319539e27b3e88c9a17e1a9f

Change-Id: I6851aa0e5ca8cecaff518609c14cd528bca95ade