Updated openstack/openstack Project: openstack/glance  967303d5e9b8e20c12ed7ca1781f7d6213c93cca
Update deprecated show_multiple_locations helptext

This option will be mentioned in an upcoming OSSN.  I think it will
be confusing to operators if the option is removed now, and it will
also be confusing if the text says "will be removed in the Ocata
release".  This patch changes the text to say "will be removed in
the Pike release or later".

Corresponding release note has been added.

Co-Authored-By: Brian Rosmaita <brian.rosmaita@rackspace.com>
Co-Authored-By: Nikhil Komawar <nik.komawar@gmail.com>

Change-Id: Ib599afaee8f48f141be125a4016aece1e40e36cf