Updated openstack/openstack Project: openstack/nova  35e4b149b61e9a9158180b83e590a26b0d1955e3
Read instances from API cell for cells v1

Recent work on cells v2 has us reading instances from compute cells
via target_cell, but that breaks the existing behavior of cells v1
because of the syncing between API cell and compute cells. When
state changes are effected in the API cell, there is a slight delay
before they are reflected in the compute cell. So, reading from the
compute cell instead of the API cell results in behavior changes.

This adds a conditional in compute/api get_all to read instances
from the API cell if cells v1 is enabled. We are already doing this
for compute/api get, and get_all was missed.

Closes-Bug: #1660747

Change-Id: I7df5c4616ef386216c7bd7efea2be68173c61be0