Updated openstack/openstack Project: openstack/designate-specs  1900a89668574fb6b0141c3158548303050c3572
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/3f4c5f18183bde7cfaef4d1bfe582355

Change-Id: I92a5cd1a51f91e90945b5ff33f181b25ff2c50d4