Updated openstack/openstack Project: openstack/barbican  4089edc1774e6018af76ad73937a943095d10e12
Adds true functional tests for db_manage script

As the db manage script will be run primarily in
production enviroments as a means to cleanup old
data. True functional tests are needed to verify
that it does not falsely delete valid information
and that it properly removes the needed informaion

This is the start of any functional tests that are
needed for our cmd support scripts.

Change-Id: Ib6c9309c4ee95d10e124869a31590cfe983e13ec