Updated openstack/openstack Project: openstack/ceilometer  b7a9f13047b8760ace3ebf8f918d9be691351a0d
Do not use non-UUID resource ID in Aodh+Gnocchi tests

The non-UUID resource id are getting to be unique per user, and since the user
creating the alarm (gabbi) and the user checking for resource existence (aodh)
are different, the string->UUID conversion gives different result.

Using an UUID instead of a string should fix that.

Change-Id: I89cb079325e327541a444559217015f38029a506