Updated openstack/openstack Project: openstack/keystone  3ace96b96f41ce94e8c5a762a7bedfb67dee5c70
Add validation that token method isn't needed in MFARules

Add validation test that token method is not needed explicitly
in the MFA rules to be utilized. Token should work as all previous
methods (prior to the rescope) are included in the token's methods
when token is the authmethod handler.

This change makes some minor restructureing to the test class to
ensure sharing of code when handling TOTP credentials.

Change-Id: Ib1d9f0652600b5eb488ae19fc7671c8d69327c11
bp: per-user-auth-plugin-reqs