Updated openstack/openstack Project: openstack/keystone  8938ec449235cc64bb70d6ec8e38c2b78742bb3d
Remove de-dupe for MFA Rule parsing.

The de-duplication was over optimisation and not of value it is
less expensive to have 1 or 2 (at most) rules loaded form storage
(which is already protected against having multiple rules stored
both via json schema and the data storage bits itself) than to
join the strings and sort them just to eliminate another iteration.

Change-Id: I88b3ab0e956e32be91f87d85cb4e19069dd3d08c