Updated openstack/openstack Project: openstack/nova  89c1c6ff619ffaac0e81fbcd15134e1873cf5437
Allow scheduler to run cell host discovery periodically

Since cellsv2 has required host mappings to be present, we have lost
our automatic host registration behavior. This is a conscious choice,
given that we do not want to have upcalls from the cells to the API
database. This adds a periodic task to have the scheduler do the
discovery routine, which for small deployments may be an acceptable
amount of background overhead in exchange for the automatic behavior.

There is also probably some amount of performance improvement that
can be added to the host discovery routine to make it less of an
issue. However, just generalizing the existing routine and letting
it run periodically gives us some coverage, given where we are in
the current cycle.

Related to blueprint cells-scheduler-integration

Change-Id: Iab5f172cdef35645bae56b9b384e032f3160e826