Updated openstack/openstack Project: openstack/tripleo-heat-templates  8a44985ed0d05643f53ba3becdaf2a5f97a6d389
Don't run ceilometer-upgrade via upgrade_tasks

This needs to be run by puppet or ansible runs it as root and the
later run by puppet fails due to permissions on the logfile.

Probably we need to remove the *sync calls for most services to
avoid similar issues, now that we're running puppet as part of the
pre-converge upgrade process but that will be done in another patch.

Change-Id: I808db2c175325a25058226842684558ea06fb5c5
Partially-Implements: blueprint overcloud-upgrades-per-service