Updated openstack/openstack Project: openstack-infra/tripleo-ci  234b51a39dc2605c090afa08e23fb09b0208bdd9
Fix image size metric retrieval

I'm not sure what the purpose of this sed was, but all it's doing
right now is cutting off the last digit of the image size.  So an
image of 1152 MB is being reported as 115.

My best guess is that this was designed to remove a newline from
the output, but in my local testing there is no newline.  Maybe
something changed in the output from du/cut.

Change-Id: I225daf1e3abfe627713b1f54fae902236b5c321f