Updated openstack/openstack Project: openstack-infra/project-config  f8b6567bb092f22fe24ac3a8b89486a4e9838922
Run containers job both with regex and in experimental

When we introduced regexps for contaners job in tht project,
we lost any option to run containers job on demand for other
projects. This patch leaves -nv job for regexps and checking
THT repo and job without suffix will run in experimental queue.

Change-Id: I643d291c63855c85e06bc4d111005487304b2948