Updated openstack/openstack Project: openstack-infra/project-config  91e33aeb1113aa0b2a80fc3f65e825b662e54907
Create the ansible-role-ara project

The community of users around the ARA [1] project are coming up with
different Ansible roles to install and set up ARA.
To ensure a quality standard and have a home where these users can
contribute their code in a common place, we'd like to host this
project side-by-side the ARA project.

There is no launchpad for the ansible-role-ara project as we will
be using storyboard like we have been doing with ARA.

There is also no gate jobs for the time being as we figure how
we will implement and share them between the main project and this
one.
New gerrit groups are created for this project as the expectation
is that ara-core will be a member of ara-ansible-role-core but the
inverse might not be true.

[1]: https://git.openstack.org/cgit/openstack/ara

Change-Id: I50f83225525fba07db1c8eb940bcf642de2accbf