Updated openstack/openstack Project: openstack-infra/project-config  b917cfe565ef24e28fea618c1884037bbf78561e
Ignore node_modules when checking for po files

The translation proposal job for tripleo-ui is currently failing due to
errors in po files under the node_modules directory. However
'node_modules' is only used by nodejs in order to store dependencies,
it doesn't need to be traversed and should be blacklisted.

Change-Id: Ibec1521ce222b559ddb84f61c7589bc1400c1521