Updated openstack/openstack Project: openstack/tripleo-heat-templates  2fa4a7148d125836b9f4ee470717e563022209f5
Don't run yum_update.sh inside docker

For now, don't run anything in yum_update.sh when it is run from
inside the heat-agents container. A mechanism for doing a yum update
on the host can be worked out later, but for now a yum update should
never be run inside a container.

Change-Id: I73d37578f8b2dc9c3029b968b1ef74ef4894100a