Updated openstack/openstack Project: openstack-infra/shade  24e62f8d338cb162285d1078d8a5d274156180a0
Stop spamming logs with unreachable address message

We emit a helpful debug warning when we have to pick from amongst random
addresses and we couldn't choose one based on being able to reach it.
However, there are a number of addresses for a server that EXPECT to not
be reachable. We call them private addresses. Or maybe non-routable. Or
something fancier and full of more networking acronyms that I'm not cool
enough to know.

In any case, only emit the warning when we expect to be able to route to
the address and we still can't figure it out.

Change-Id: I72d6f4745fe9f4169671bc933e92871358cebbf2