Updated openstack/openstack Project: openstack/nova  60ec9cbf4b14577e4359f7e81542e6dc744d32d7
doc: add upgrade notes to the placement devref

This just adds some more detailed docs for upgrading
to Ocata and notable things for the Placement service
and related components as part of that upgrade.

The Ocata release notes will have some of this
information too, but I think it's good to also have
a single place for some of this, like an install/upgrade
guide for Placement.

Change-Id: Id33b9c852bb36ae80a6f29ae33d963a78cbff193
Depends-On: Ie12acb76ec5affba536c3c45fbb6de35d64aea1b