Updated openstack/openstack Project: openstack/horizon  7d5b6ebca68f001212f789958a617a9d20379946
Move Security Groups into its own panel

This patch moves the Security Groups tab from the Access and Security
panel into its own panel under the Network panel group. As this is the
last tab in Access and Security, that panel is also removed by this
patch.

Change-Id: Id29c7ce635d46383742aec140def265d4b249aa5
Implements: blueprint reorganise-access-and-security