Updated openstack/openstack Project: openstack/manila  16e2debf6969db8bb8f59b7c452fff374563ea14
HNAS: avoid mismatch access level for managed shares

A CIFS managed share can contain some rules added in backend, and when
we try to add an already existent rule through manila, but with
different access level, the driver will return success with message
"user already allowed". So, manila will show an access level for this
rule that is different from the one in backend. This behaviour does
not apply to NFS shares.

In this case, it will call an update access rule for CIFS shares to
assure that rules in backend and in manila are always with same access
level.

Change-Id: I3c97ff7a5fee241685d54ac4cf7269ed55f1f602
Closes-bug: #1660401