Updated openstack/openstack Project: openstack/cinder  c2404d71d0dec5ad181ce3dc344913c610ae8aa5
Make get_capabilities look for clusters

Currently get_capabilities accepts only host as a service identifier and
fails when it cannot find a service with such host property. In A/A
deployment, it should accept cluster_name as well, because
get_capabilities addresses backend and not individual host. This commit
fixes this up by making the call seek for available services using both
host and cluster_name properties.

Although it's an API change, it doesn't need a new microversion, because
it falls under "Did we silently fail to do what is asked?" category.

Change-Id: If7bca131c84fc997da8ad277eccd134e8ca4b316
Closes-Bug: 1660990