Updated openstack/openstack Project: openstack/ironic  ac2b1a48ef5420e9c71a665e7157dc6e55016a16
Remove iSCSI deploy support for IPA Mitaka

The Newton (6.0.0) release included a fix [1] to support
iSCSI deploys to IPA ramdisks < 1.3 (Mitaka). These older
versions of IPA did not handle wiping disk metadata or
customized portal ports.

We are in the Ocata cycle and ironic no longer needs to
support IPA Mitaka or earlier, so we can remove this fix.

[1] Commit ba83fb1bb45b63e056c87c3bbf42cabaade7865c

Change-Id: Ib97344d96a938a550ad2c61e14d793d0f56c6a89
Fixes-Bug: #1661096