Updated openstack/openstack Project: openstack/cinder  e605f606ff2896dc7a8adb6bc03486ec3012a12d
Convert ignoreskippedcluster check to be static

Using LVM.supports_XXX will check for the existence of the
property method, rather than attempting to retrieve the property.

As the caller is a static method, we can't actually use a property
here, so fix this to use static methods and cache on the class.
Fixes I89ffea86d0951fe4c80783a612b6cde76c8838b4.

Change-Id: I7294ccdf9c0c5858d949b01ad55763ad3129d8ec
Closes-bug: 1661144