Updated openstack/openstack Project: openstack-infra/tripleo-ci  3d1d3ccaf23e2a4bb121aead0c98485e68e2453f
Enable Debug: True on OVB jobs

Debug should be enabled in CI, and there is never enough logs to help in
debugging and solving complex problems.
We should not use the default value in TripleO and override it in CI to
let our developers debugging correctly.

Note: it's already enabled for multinode jobs.

Change-Id: I219214c8a77f35fcd41414b3d83b3ebb84c4cb0f