Updated openstack/openstack Project: openstack/horizon  b784015898f62069028349dc799599005d029fbb
Pass the project domain ID to novaclient

When the nova client creates a new session, knowing only the ID and
project of a token but not the project's domain, it calls keystoneauth
with None as the project domain. Keystoneauth takes this to mean
novaclient wants to authenticate with the Identity v2.0 API, which
fails if the user is a federated user. This patch grabs the project
domain ID from the request and passes it along to novaclient so that it
can properly create a v3 Token session.

Change-Id: I7f00a9fefd64c453c40b08fbe235ac7e42153050
Depends-on: I18a9d42906cb2116903600d47880ebdfff1e1ef9
Partial-bug: #1660436