Updated openstack/openstack Project: openstack/nova  ed27145a0b6c1afb57436c66ec55d30c1623808b
Stop swap allocations being wrong due to MB vs GB

Swap is in MB, but allocations for disk are in GB.

We totally should claim disk in GB, for now lets just round up the swap
allocation to the next GB. While this is wasteful, its the only safe
answer to ensure you don't over commit resources on the node.

Updated the test so the swap is 1023MB, after rounding up this should
claim the same 1GB extra space for swap.

Closes-Bug: #1659266

Change-Id: If50eab870b2c50f4055668143780574e1350a438