Updated openstack/openstack Project: openstack/nova  63cfd90f035b5ab03bcaa682cd2d640811abfed6
nova-status: relax the resource providers check

As of 4660333d0d97d8e00cf290ea1d4ed932f5edc1dc the filter
scheduler will fallback to not using the placement service
if the minimum nova-compute service version in the deployment
is not new enough to ensure the computes are reporting into
the placement service.

This means we need to relax our check for when there are no
resource providers but there are compute nodes, since the filter
scheduler will not fail on that in Ocata. We intend on making
that a hard failure in Pike, at which time we'll need to adjust
nova-status again.

Change-Id: I1e4dae17cf9d1336bf0ca72948b135b02434ba15
Closes-Bug: #1660484