Updated openstack/openstack Project: openstack/neutron  1a5fbc862bab53f30e7292189cdbb637324fe938
Fix iptables rules for Prefix Delegated subnets

Make sure the correct iptables rule is added when the router gets
an interface on a PD-enabled subnet. This will allow traffic on PD
subnets to reach the external network.

Includes a unit test for the new function, and modifies an
existing test to verify the adding and removal of the rule.

Change-Id: I42f8f42995e9809e5bda2b29726f7244c052ca1c
Closes-Bug: #1570122