Updated openstack/openstack Project: openstack/nova  6f9c845f749f59c39ef962cc28cfe5fa5745222c
libvirt: workaround findmnt behaviour change

findmnt as provided by util-linux-2.23.2-33 now returns 0 when --source
is already mounted on the host under a different mountpoint [1]. This causes
issues in environments where cinder-volume may have already mounted a
given NFS share.

This change works around this behaviour change by using os.path.ismount
to confirm if the target is already a mountpoint and if a source is
provided also checking /proc/mounts to confirm that it is mounted at the
target location.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1405238

Closes-Bug: #1656774
Change-Id: Id64cc77237bf770a9a75bbf666a54a604ace558c