Updated openstack/openstack Project: openstack-dev/grenade  b9467baeb4aa51e7134d1d7ee56d8414a888f6e3
Making is_ubuntu func available to fetch_devstacks

When the fetch_devstacks function runs as part of grenade.sh
it tries to install devstack tools. As part of the devstack tools
installation, python3 gets installed through a function called
install_python3 which uses a function called is_ubuntu to
determine if the system is ubuntu. However this is_ubuntu
function is not available at that moment, so the script fails.

This commit sources functions-common from the devstack directory
which includes is_ubuntu so the grenade.sh script does not break.

Change-Id: I92b9f731e99df0840be296bfa8ce6a70e02a1f91
Closes-bug: 1659081