Updated openstack/openstack Project: openstack/nova  245c54796f790cb70703f069bc429a8b9c5e0e49
Delete a compute node's resource provider when node is deleted

Currently when a compute node is deleted, its record in the cell DB is
deleted, but its representation as a resource provider in the placement
service remains, along with any inventory and allocations. This could
cause the placement engine to return that provider record, even though
the compute node no longer exists. And since the periodic "healing" by
the resource tracker only updates compute node resources for records in
the compute_nodes table, these old records are never removed.

This patch adds a call to delete the resource provider when the compute
node is deleted. It also adds a method to the scheduler report client
to make these calls to the placement API.

Partial-Bug: #1661258
Closes-Bug: #1661014

Change-Id: I6098d186d05ff8b9a568e23f860295a7bc2e6447