Updated openstack/openstack Project: openstack/keystone  9bb57c1eb7f295d1c7d37f880f5ed99bbee766a4
Update endpoint api for optional region_id

As per the bug, 'name' and 'nme' are not part of the 'endpoint'
table and were being assigned to the 'extras' column. This is
why they are not being validated.

The endpoint docs also show that `region_id` is not optional, even
though it is. This updates the docs to reflect optional `region_id`.

Closes-Bug: #1579014
Change-Id: I085b75c59767eb96b3bdfe3b887e5e2639122a34