Updated openstack/openstack Project: openstack/tripleo-heat-templates  340a3dffbb423322ef28a99a893771bed12e53aa
net-config-multinode: make controlplane int idempotent

When doing a stack-update, it will try to create the control plane
interface again.
Add this conditional so the interface is not created if already exist.

Note: this code has been taken from tripleo-ci and is consistent with
how multinode jobs are currently tested.

Co-Authored-By: James Slagle <jslagle@redhat.com>
Co-Authored-By: Steven Hardy <shardy@redhat.com>
Co-Authored-By: Mathieu Bultel <mbultel@redhat.com>

Change-Id: I773fdf5359cead6961b595e3c8192b02406452b7
Related-Bug: #1661412