Updated openstack/openstack Project: openstack/heat  3e58a8b555d5f9e4cea9f75a150824d95caa96fd
Log stack status changes in unit tests

We no longer always persist status changes to the database in
Stack.state_set() - in most cases we wait until the stack lock is released
when existing the thread, so as to be able to make the change cheaply and
atomically. However, in the unit tests we don't generally have a lock and
operations are not generally run in a thread anyway. Since the logging is
close to the database storage call, this means that there is no logging
when a stack fails in a unit test. This makes debugging failing tests a
giant pain, as you can never see what happened.

This moves the INFO log back to the state_set() method itself. To replace
the INFO log, there's also a DEBUG log so we can continue to see when the
state actually gets persisted.

Change-Id: I781c403ac07b1c9b25ac815600b37be161103ce7