Updated openstack/openstack Project: openstack/cinder  9399d3009982b86da8ace4486844023348eeb3e9
VMAX driver - sync sv issue

We have a function which returns a list of all storage
synchronized instances (pairs of volumes in a replication
relationship) associated with a particular volume. This is
done to get any snapshots associated with a volume before
deletion, but it fails to differentiate between local
replication and remote replication volume pairs. This can
cause issues when deleting a replicated volume. This patch
adds a check to ensure the target volume's storage system
is the same as the source volume's storage system.

Change-Id: I9ae765c5da35fe2fb0b4001439070f1c6d32e095
Closes-Bug: #1660368